requirements.txt
absl-py==2.3.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
async-timeout==5.0.1
attrs==25.4.0
autograd==1.8.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cmake==3.25.0
colorama==0.4.6
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
e2cnn==0.2.3
escnn==1.0.11
faiss-cpu==1.7.2
faiss-gpu==1.7.2
filelock==3.14.0
fire==0.7.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.41.1
google-auth-oauthlib==0.4.6
grpcio==1.75.1
h5py==3.14.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.2
kiwisolver==1.4.7
lazy_loader==0.4
lie_learn==0.0.2
lightning-utilities==0.15.2
lit==15.0.7
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.5.3
mdurl==0.1.2
mmcv==1.7.2
model-index==0.1.11
mpmath==1.3.0
multidict==6.7.0
networkx==3.2.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.3.1
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.3
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==1.5.1
Pillow==9.2.0
platformdirs==4.4.0
prettytable==3.5.0
propcache==0.4.1
protobuf==3.20.3
py3nj==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
pymanopt==2.2.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.0
pytorch-metric-learning==1.6.3
pytz==2023.4
PyYAML==6.0.3
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.2.1
scipy==1.9.3
seaborn==0.13.2
shapely==2.0.7
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.4
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==0.6.12
tomli==2.3.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==0.11.1
torchvision==0.15.2+cu118
tqdm==4.65.2
triton==2.0.0
typing_extensions==4.15.0
urllib3==1.26.20
wcwidth==0.2.14
Werkzeug==3.1.3
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
